Add alpha script and fix detection.
